+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Convert augury config from ini style to yaml (breaking change)
 - Add options for filtering out tweets (retweets, replies, links)
+- Add option to show twitter user name as an attribution
 
 ## 0.3.0 (2015-08-20)